pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-research-app"
version = "1.0.0"
description = "Find nearby places and extract business contact information using Google Maps API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Luwalekeah", email = "your-email@example.com"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
keywords = ["google-maps", "places-api", "market-research", "lead-generation", "business-contacts"]

dependencies = [
    "googlemaps>=4.10.0",
    "pandas>=1.5.0",
    "geopy>=2.4.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "streamlit>=1.31.0",
    "streamlit-folium>=0.18.0",
    "folium>=0.15.0",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.scripts]
market-research = "cli:main"

[project.urls]
Homepage = "https://github.com/Luwalekeah/market-research-app"
Repository = "https://github.com/Luwalekeah/market-research-app"
Issues = "https://github.com/Luwalekeah/market-research-app/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src --cov-report=term-missing"
